+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 2019/5/24
+
+- **`react/autobind-light`**
+  - action: `add`
+  - result: `release`
+
 ## 2019/5/17
 
 - **`redux/Action`**